removing titus-ext/elasticsearch module, refactoring taskspublisher classes into a new subpackage
