+++ build.gradle
@@ -87,7 +87,6 @@ subprojects {
         okHttpVersion = '3.8.0'
         cassandraDriverVersion = '3.3.+'
         commonsCliVersion = '1.3.+'
-        elasticsearchVersion = '2.4.2'
         caffeineVersion = '2.6.+'
         rxJavaInteropVersion = '0.13.+'
