+++ titus-supplementary-component/tasks-publisher/src/main/java/com/netflix/titus/supplementary/taskspublisher/config/TasksPublisherConfiguration.java
@@ -21,10 +21,10 @@ import java.util.concurrent.TimeUnit;
 import com.netflix.spectator.api.DefaultRegistry;
 import com.netflix.titus.grpc.protogen.JobManagementServiceGrpc;
 import com.netflix.titus.grpc.protogen.JobManagementServiceGrpc.JobManagementServiceStub;
-import com.netflix.titus.supplementary.taskspublisher.DefaultEsWebClientFactory;
-import com.netflix.titus.supplementary.taskspublisher.EsClient;
-import com.netflix.titus.supplementary.taskspublisher.EsClientHttp;
-import com.netflix.titus.supplementary.taskspublisher.EsWebClientFactory;
+import com.netflix.titus.supplementary.taskspublisher.es.DefaultEsWebClientFactory;
+import com.netflix.titus.supplementary.taskspublisher.es.EsClient;
+import com.netflix.titus.supplementary.taskspublisher.es.EsClientHttp;
+import com.netflix.titus.supplementary.taskspublisher.es.EsWebClientFactory;
 import com.netflix.titus.supplementary.taskspublisher.TasksPublisherCtrl;
 import com.netflix.titus.supplementary.taskspublisher.TitusClient;
 import com.netflix.titus.supplementary.taskspublisher.TitusClientImpl;
