+++ titus-supplementary-component/tasks-publisher/src/main/java/com/netflix/titus/supplementary/taskspublisher/es/EsClient.java
@@ -13,12 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.netflix.titus.supplementary.taskspublisher;
+package com.netflix.titus.supplementary.taskspublisher.es;
 
 
 import java.util.List;
 
-import com.netflix.titus.ext.elasticsearch.TaskDocument;
 import reactor.core.publisher.Mono;
 
 public interface EsClient {
@@ -142,7 +141,7 @@ public interface EsClient {
         }
     }
 
-    public static class IndexHeaderLine {
+    class IndexHeaderLine {
         private IndexHeader index;
 
         public IndexHeader getIndex() {
@@ -154,7 +153,7 @@ public interface EsClient {
         }
     }
 
-    public static class IndexHeader {
+    class IndexHeader {
         private String _index;
         private String _type;
         private String _id;
